Writing to files update

+++ helpers.cpp
@@ -6,7 +6,7 @@ void delayByLoop(uint32_t nCount){
 }
 
 void reverse(char s[]){
-	unsigned short i, j;
+	unsigned int i, j;
 	char c;
 	for (i = 0, j = strlen(s)-1; i<j; i++, j--) {
 		c = s[i];
